+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v6.4.0
 ------
 
+* Add RA and dec information on dome summary table components and a few other improvements. `<https://github.com/lsst-ts/LOVE-frontend/pull/675>`_
 * Add M1M3HardpointsDataTable component. `<https://github.com/lsst-ts/LOVE-frontend/pull/674>`_
 * Add M1M3Compact and M2Compact components. `<https://github.com/lsst-ts/LOVE-frontend/pull/673>`_